[FIX] EplNmtMnu: Always forward kEplNmtNodeEventFound to application
Previously, the kEplNmtNodeEventFound-Event (IdentResponse received) was
only forwarded to application at the very first time, but not after an
NMT ResetConfig or ResetNode command. This changes now, because the
application might be interested in all IdentResponses received.

If necessary in the future, the API could be extended to inform the
application about the internal state/reason for the IdentResponse.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>