[FIX] ProcessImage: Use proper ref counters for Linux kernel
This fixes a long standing kernel crash that happened when user space
application (e.g. console or Qt demo) is terminated via CTRL-C.

In case of CTRL-C the wait_for_completion_interruptible() call inside
EplApiProcessImageExchange() exited and the completion structure is freed.
Later EplApiProcessImageFree() accesses this structure with
EplApiProcessImageSignalCompletion() again, because the copy job is still
in the queue. The write access to already freed memory crashed the kernel.

This is solved now by using reference counters (see kref.h) for the
internal completion structure. So this structure is freed after both
EplApiProcessImageSignalCompletion() and EplApiProcessImageExchange() have
finished their job.

Signed-off-by: Daniel Krueger <daniel.krueger@systec-electronic.com>